+++ baybe/optimizers/__init__.py
@@ -1,0 +1,9 @@
+"""Optimizers."""
+
+from baybe.optimizers.base import OptimizerProtocol
+from baybe.optimizers.continuous import ContinuousOptimizer
+
+__all__ = [
+    "ContinuousOptimizer",
+    "OptimizerProtocol",
+]