+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `optimizers` subpackage with `OptimizerProtocol` and `ContinuousOptimizer`
+
+### Removed
+- `recommend_continuous_without_cardinality_constraints` function.
+   Call `ContinuousOptimizer` instead.
+
+### Deprecations
+- `BotorchRecommender` class. Use `BayesianRecommender` instead.
+
 ## [0.15.0] - 2026-06-11
 ### Breaking Changes
 - `GaussianProcessSurrogate` no longer automatically adds a task kernel in multi-task